reinstalled C# dev kit, generated assets for debugger, reorganized react folders
