+++ client-app/src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './app/layout/styles.css'
 import 'semantic-ui-css/semantic.min.css'
-import App from './App.tsx'
+import App from './app/layout/App.tsx'
 
 createRoot(document.getElementById('root')!).render(
     <StrictMode>
